# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/users/{userId}/deposits/fiat": {
      "get": {
        "operationId": "AdminDepositsController_listFiatDeposits",
        "summary": "List user fiat deposits (admin only)",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max results (1-100)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Skip N results for pagination",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Start date (ISO 8601)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "End date (ISO 8601)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns fiat deposits",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": [
                      {
                        "id": "dd0e8400-e29b-41d4-a716-446655440030",
                        "userId": "550e8400-e29b-41d4-a716-446655440000",
                        "fiatCurrency": "USD",
                        "amount": "500",
                        "fee": "2.5",
                        "status": "completed",
                        "createdAt": "2026-04-09T12:00:00.000Z",
                        "completedAt": "2026-04-09T12:00:00.000Z"
                      }
                    ],
                    "total": 1,
                    "meta": {
                      "total": 1,
                      "limit": 20,
                      "offset": 0
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Admin access required",
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "User not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-deposits"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/users/{userId}/deposits/fiat/{id}": {
      "get": {
        "operationId": "AdminDepositsController_getFiatDeposit",
        "summary": "Get user fiat deposit by ID (admin only)",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Deposit ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns fiat deposit",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "dd0e8400-e29b-41d4-a716-446655440030",
                    "userId": "550e8400-e29b-41d4-a716-446655440000",
                    "fiatCurrency": "USD",
                    "amount": "500",
                    "fee": "2.5",
                    "status": "completed",
                    "createdAt": "2026-04-09T12:00:00.000Z",
                    "completedAt": "2026-04-09T12:00:00.000Z"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Admin access required",
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User or deposit not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "User not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-deposits"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/users/{userId}/deposits/crypto": {
      "get": {
        "operationId": "AdminDepositsController_listCryptoDeposits",
        "summary": "List user crypto deposits (admin only)",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max results (1-100)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Skip N results for pagination",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns crypto deposits",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": [
                      {
                        "id": "ff0e8400-e29b-41d4-a716-446655440050",
                        "userId": "550e8400-e29b-41d4-a716-446655440000",
                        "amount": "0.05",
                        "walletAddress": "bc1qabc",
                        "status": "completed",
                        "createdAt": "2026-04-09T12:00:00.000Z",
                        "symbol": "BTC"
                      }
                    ],
                    "total": 1,
                    "meta": {
                      "total": 1,
                      "limit": 20,
                      "offset": 0
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Admin access required",
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "User not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-deposits"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/users/{userId}/deposits/crypto/{id}": {
      "get": {
        "operationId": "AdminDepositsController_getCryptoDeposit",
        "summary": "Get user crypto deposit by ID (admin only)",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Deposit ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns crypto deposit",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "ff0e8400-e29b-41d4-a716-446655440050",
                    "userId": "550e8400-e29b-41d4-a716-446655440000",
                    "amount": "0.05",
                    "walletAddress": "bc1qabc",
                    "status": "completed",
                    "createdAt": "2026-04-09T12:00:00.000Z",
                    "symbol": "BTC"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Admin access required",
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User or deposit not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "User not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-deposits"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Test Exchange API",
    "description": "Crypto exchange training platform API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}